refactor: divide toast component file
